迁移到WSL，并适配新的gcc版本

+++ src/main.c
@@ -41,17 +41,17 @@ int main(int argc, char* argv[]){
     clock_t begin, end;
     double time_cost;
     uint64_t INST_NUM = 100;
-    int ips = 0;
+    double ips = 0;
     print_localtime();
     begin = clock();
     cpu_start(INST_NUM);
     end = clock();
     time_cost = (double)(end-begin)/CLOCKS_PER_SEC;
-    ips = INST_NUM/(int)(time_cost);
+    ips = (double)(INST_NUM)/time_cost;
     printf("--------------------------------\n");
-    printf("%d Instructions Simulated!\n",INST_NUM);
+    printf("%lu Instructions Simulated!\n",INST_NUM);
     printf("Time Cost: %f\n",time_cost);
-    printf("Instruction Number Per Second: %d\n",ips);
+    printf("Instruction Number Per Second: %f\n",ips);
     printf("--------------------------------\n");
 
     return 0;
